Is the building allowed base on zoning constraints?

zr_check_constraints() compares the building characteristics against many of the main constraints in the .zoning file. It returns a one-row data frame with a column for each constraint it checked against.

Usage

zr_check_constraints(
  vars,
  zoning_req,
  checks = c("far", "fl_area", "fl_area_first", "fl_area_top", "footprint", "height",
    "height_eave", "lot_cov_bldg", "lot_size", "parking_enclosed", "stories",
    "unit_0bed", "unit_1bed", "unit_2bed", "unit_3bed", "unit_4bed", "unit_density",
    "unit_pct_0bed", "unit_pct_1bed", "unit_pct_2bed", "unit_pct_3bed", "unit_pct_4bed",
    "total_units", "unit_size_avg")
)

Arguments

vars

the result from the zr_get_variables() function. If this data frame is supplied, bldg_data, parcel_data, and zoning_data are not needed.

zoning_req

The output of zr_get_zoning_req(). A data frame with constraint values

checks

A list of all the checks that should take place. The default is every check possible. Note, if a zoning file doesn't have zoning info for one of the constraints listed in the checks variable, then it is assumed that building characteristic is allowed.