Find where a building is allowed to be built

zr_run_zoning_checks() checks the building information against all the zoning constraints to see which parcels will allow the building.

Usage

zr_run_zoning_checks(
  bldg_file,
  parcel_files,
  zoning_files,
  detailed_check = FALSE,
  print_checkpoints = TRUE,
  checks = possible_checks,
  save_to = NULL
)

Arguments

bldg_file

The path to the OZFS *.bldg

parcel_files

The path to an OZFS .parcel file, or the path to a folder containing OZFS .parcel files

zoning_files

The path to an OZFS .zoning file, or the path to a folder containing OZFS .zoning files

detailed_check

When TRUE, every parcel passes through each check no matter the result, and it takes more time. When FALSE, subsequent checks are skipped as soon as one check reads FALSE

print_checkpoints

When TRUE, runtimes and other info will be printed at certain points throughout the function.

checks

A list of all the checks that should take place. The default is to check for every constraint possible in the OZFS. These constraints can found in the package data possible_checks. Note, if a zoning file doesn't have zoning info for one of the constraints listed in the checks variable, then it is assumed that building characteristic is allowed.

save_to

The path or folder directory where you want to save the results as a geojson file.

Value

A simple features data frame with geometry for the centroid of each parcel and columns to show info on allowance of the building.

Zoning Checks

The checks that can take place in zr_run_zoning_checks() are included in the table below. Most checks are not required and only checked if specified by the user in the checks argument. All optional checks possible are stored in the possible_checks data which is the default value of the checks argument. A few checks are built in to the function and, therefore, not a part of the checks argument.

The resulting data frame of zr_run_zoning_checks() will have a "reason" column with the names of any checks that caused a FALSE or a MAYBE building allowance. The table below specifies what each check means when it is part of the "reason" column.

Possible Checks	Type	Reason Column Usage

PD_dist	built in	FALSE returned because parcel is in a planned development district

PD_overlay	built in	FALSE returned because parcel is in a planned development overlay district

side_lbl	built in	MAYBE returned because the parcel sides were not labeled and check_fit was not run

res_type	user input	Residential type check returned FALSE or MAYBE

far	user input	Floor Area Ratio check returned FALSE or MAYBE

fl_area	user input	Floor area check returned FALSE or MAYBE

fl_area_first	user input	Area of first floor check returned FALSE or MAYBE

fl_area_top	user input	Area of top floor check returned FALSE or MAYBE

footprint	user input	Footprint check returned FALSE or MAYBE

height	user input	Height check returned FALSE or MAYBE

height_eave	user input	Height eave check returned FALSE or MAYBE

lot_cov_bldg	user input	Lot coverage check returned FALSE or MAYBE

lot_area	user input	Lot area check returned FALSE or MAYBE

parking_enclosed	user input	Enclosed parking check returned FALSE or MAYBE

stories	user input	Stories check returned FALSE or MAYBE

unit_0bed	user input	0-bed unit quantity check returned FALSE or MAYBE

unit_1bed	user input	1-bed unit quantity check returned FALSE or MAYBE

unit_2bed	user input	2-bed unit quantity check returned FALSE or MAYBE

unit_3bed	user input	3-bed unit quantity check returned FALSE or MAYBE

unit_4bed	user input	4plus-bed unit quantity check returned FALSE or MAYBE

unit_density	user input	Unit density check returned FALSE or MAYBE

unit_pct_0bed	user input	Percent 0-bed units check returned FALSE or MAYBE

unit_pct_1bed	user input	Percent 1-bed units check returned FALSE or MAYBE

unit_pct_2bed	user input	Percent 2-bed units check returned FALSE or MAYBE

unit_pct_3bed	user input	Percent 3-bed units check returned FALSE or MAYBE

unit_pct_4bed	user input	Percent 4plus-bed units check returned FALSE or MAYBE

total_units	user input	Total units check returned FALSE or MAYBE

unit_size_avg	user input	Average unit size check returned FALSE or MAYBE

unit_size	user input	Unit size check returned FALSE or MAYBE

bldg_fit	user input	Building fit check returned FALSE or MAYBE

overlay	user input	MAYBE returned because parcel is in an overlay district